feature: introduce esp_lib_idf_helpers.h
it provides macro definitions to identify build target and SDK version.
this header is a temporary workaround until ESP_IDF_VERSION_VAL is
available in all the supported SDKs. ESP_IDF_VERSION_VAL is only
available in the latest (4.x) esp-idf for ESP32.

Supersedes a hackish workaround suggested in issue #42.

https://github.com/UncleRus/esp-idf-lib/issues/42