implement visual indication of data level
in the original code, GPIO matrix is used to turn onboard LED on/off.
ESP8266 does not have GPIO matrix. turn the LED in a task, which is not
very accurate, but enough to test the example.

also, implement a simple task that tell users that the calibration
should be done.