ci: CI build with GitHub Actions
As the CI build in Travis CI is slow, and its scalability is limited by
its small number of parallel jobs, the CI process was ported to GitHub
Actions.

Major improvements include:

- support esp-idf 4.x versions and 3.1.7
- faster build (22 minutes 37 seconds with 10 build jobs, including
  esp-idf 4.x versions, and 3.1.7).
- sharing caches between different examples (the SDK code is compiled
  just once).
- the cleaner implementation (but more duplicates) allows maintainers to
  modify one build job without affecting other jobs.

A drawback of GitHub Actions: When a job in a matrix fails, other jobs
in the matrix will be terminated. When your change breaks version 3.1,
you cannot see if it also brakes version 3.3 because the jobs are in
same job matrix. When 3.1 build fails, GitHub immediately terminates
3.3 job. However, ESP8266 job will not be terminated because ESP8266 job
is defined in another job.

Possible further optimizations:

- cache downloaded files (as it's already fast, the improvement will be
  minor)
- cache ccache directory, which will improve cache hit rate in
  subsequent builds in the branch