ci: introduce skip-duplicate-actions GitHub Actions
to remove duplicated jobs from the CI pipeline. the original
implementation did the followings:

- when a change is pushed, or merged, to a branch, run the workflow
- when a PR is created, run the workflow

the result is, the workflow runs all jobs on the same contents twice
in a typical CI pipeline, "create a feature branch, make a PR, and merge
the PR", once when you push changes, and another when you create a PR.

this commit attempts to remove the duplication: if the change is tested
in a future branch, skip the jobs in PR. also, skip the jobs when the
change includes documentation changes only.