ci: add `all_build` job to build workflow
this job is a meta job that requires all other build jobs. "Require
branches to be up to date before merging" in branch protection setting,
one needs to choose at least one test to pass. However, when the check
name is renamed, Pull Request is blocked because the new name is not
listed in the setting, and the repository admin needs to change the
setting everytime a test is renamed, added, or removed. This job can be
used for the test to pass.