meta: remove skip form tests to fix a bug
the bug is: when a component has two copyrights, and one has author,
then thr other one is not tested to have a valid Person. this is because
`skip` skips the example when one caused `unless copyright.author?` to
be true. use `select` to ignore copyright without author.

while here:

use more `describe` to clarify what is being tested