ci: add esp32c3 build to CI
fixes #281

esp32c3 support is not fully confirmed on real hardware. however, as the
support in esp-idf is a very recent change, there should be no
fundamental incompatibilities. some defaults in components might be
incompatibile, but users can override them in their code. such
incompatibilities cannot be found in CI because it's a run-time error.
we will fix them as they are found.