ci: continue jobs even when a job in matrix fails
by default, fail-fast is true. this means when one job in a matrix fails,
all other jobs in the matrix are canceled. this is not useful when you
want to know which job, or which combination of target, build_method,
and version, is failing. an example is #328.

https://github.com/UncleRus/esp-idf-lib/pull/328

see Handling failures at:
https://docs.github.com/en/actions/using-jobs/using-a-matrix-for-your-jobs#handling-failures