ci: trigger CI tests when "Label PR" completes
when a labeler.yml add labels to a PR, that does not trigger other
workflows to prevent accidental recursive workflow runs. that means,
with the old implementation, when a member of the repository manually
add labels to a PR, all CI tests run as intended. however, when GitHub
Action bot (the workflow to label PRs) add labels, the `labeled` event
type will not trigger other CI tests.

https://docs.github.com/en/actions/using-workflows/triggering-a-workflow#triggering-a-workflow-from-a-workflow

A workaround is to use Personal Access Token, as suggested the link
above. However, i do not like Personal Access Token because the token
represents a person, and the token must be updated if you follow the
advice from GitHub (they don't like a token without expiration).

instead, use workflow_run, which can invoke other workflows without
Personal Access Token. workflow_run limits the number of chained
workflows, 3 at max, which is enought for now.

https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#workflow_run